pyproject.toml
[project]
name = "csktandem_bot"
version = "0.1.0"
description = ""
authors = ["Dmitry <dmitryusin@gmail.com>"]
requires-python = ">=3.11"
dependencies = ["flask>=2.2.2", "requests>=2.28.1", "python-dotenv>=1.0.0"]
